Remove unnecessary shebangs and permissions
These modules aren't designed to be runnable, so the shebang
and executable bit shouldn't be here.

The motivation for this cleanup is that packaging guidelines for
distros (e.g. Fedora) consider the usage of "/usr/bin/env python"
undesirable. Since it's not doing anything here, it can simply be
removed.